feat(core): reorganise and PR feedback
