+++ packages/core/core/src/services/document-service/transform/fields.ts
@@ -1,0 +1,3 @@
+export const transformFields = (fields: string[]): string[] => {
+  return fields.map((field) => (field === 'id' ? 'documentId' : field));
+};
