+++ packages/core/core/src/services/document-service/transform/sort.ts
@@ -1,0 +1,6 @@
+export const transformSort = (sort: string[] | string): string[] | string => {
+  if (Array.isArray(sort)) {
+    return sort.map((item) => (item === 'id' ? 'documentId' : item));
+  }
+  return sort.replace('id', 'documentId');
+};
